+++ Sources/GroupOperation.swift
@@ -82,18 +82,15 @@ open class GroupOperation: AdvancedOperation {
   ///
   /// - Parameters:
   ///   - operations: The operations of which the `GroupOperation` is composed of.
-  ///   - exclusivityManager: An instance of `ExclusivityManager`.
   ///   - qualityOfService: The default service level to apply to operations executed using the queue.
   ///   - maxConcurrentOperationCount: The maximum number of queued operations that can execute at the same time.
   ///   - underlyingQueue: An optional DispatchQueue which defaults to nil, this parameter is set as the underlying queue of the group's own `AdvancedOperationQueue`.
   /// - Note: If the operation object has an explicit quality of service level set, that value is used instead.
   public convenience init(operations: Operation...,
-                          exclusivityManager: ExclusivityManager = .sharedInstance,
                           qualityOfService: QualityOfService = .default,
                           maxConcurrentOperationCount: Int = OperationQueue.defaultMaxConcurrentOperationCount,
                           underlyingQueue: DispatchQueue? = .none) {
     self.init(operations: operations,
-              exclusivityManager: exclusivityManager,
               qualityOfService: qualityOfService,
               maxConcurrentOperationCount: maxConcurrentOperationCount,
               underlyingQueue: underlyingQueue)
@@ -103,17 +100,15 @@ open class GroupOperation: AdvancedOperation {
   ///
   /// - Parameters:
   ///   - operations: The operations of which the `GroupOperation` is composed of.
-  ///   - exclusivityManager: An instance of `ExclusivityManager`.
   ///   - qualityOfService: The default service level to apply to operations executed using the queue.
   ///   - maxConcurrentOperationCount: The maximum number of queued operations that can execute at the same time.
   ///   - underlyingQueue: An optional DispatchQueue which defaults to nil, this parameter is set as the underlying queue of the group's own `AdvancedOperationQueue`.
   /// - Note: If the operation object has an explicit quality of service level set, that value is used instead.
   public init(operations: [Operation],
-              exclusivityManager: ExclusivityManager = .sharedInstance,
               qualityOfService: QualityOfService = .default,
               maxConcurrentOperationCount: Int = OperationQueue.defaultMaxConcurrentOperationCount,
               underlyingQueue: DispatchQueue? = .none) {
-    let queue = AdvancedOperationQueue(exclusivityManager: exclusivityManager, underlyingQueue: underlyingQueue)
+    let queue = AdvancedOperationQueue(underlyingQueue: underlyingQueue)
     queue.qualityOfService = qualityOfService
     queue.maxConcurrentOperationCount = maxConcurrentOperationCount
     queue.isSuspended = true