+++ Tests/AdvancedOperationTests/Conditions/NoCancelledDependenciesConditionTests.swift
@@ -27,8 +27,7 @@ import XCTest
 final class NoCancelledDependenciesConditionTests: XCTestCase {
 
   func testTwoLevelCondition() {
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
 
     let operation1 = SleepyAsyncOperation(interval1: 1, interval2: 1, interval3: 1)
     let operation2 = SleepyAsyncOperation(interval1: 1, interval2: 1, interval3: 1)
@@ -63,8 +62,7 @@ final class NoCancelledDependenciesConditionTests: XCTestCase {
   }
 
   func testAllOperationCancelled() {
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
 
     let expectation1 = expectation(description: "\(#function)\(#line)")
     let expectation2 = expectation(description: "\(#function)\(#line)")
@@ -116,8 +114,7 @@ final class NoCancelledDependenciesConditionTests: XCTestCase {
   }
 
   func testWithNoFailedDependeciesCondition() {
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
 
     let operation1 = NotExecutableOperation()
     let operation2 = SleepyAsyncOperation()