+++ Tests/AdvancedOperationTests/InjectionTests.swift
@@ -43,8 +43,7 @@ class InjectionTests: XCTestCase {
     let operation2 = StringToIntOperation()
     operation1.input = 10
     let adapterOperation = AdvancedOperation.injectOperation(operation1, into: operation2)
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
     queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
 
     XCTAssertEqual(operation2.output, 10)
@@ -55,8 +54,7 @@ class InjectionTests: XCTestCase {
     let operation2 = StringToIntOperation()
     operation1.input = 10
     let adapterOperation = operation1.inject(into: operation2)
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
     queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
 
     XCTAssertEqual(operation2.output, 10)
@@ -73,8 +71,7 @@ class InjectionTests: XCTestCase {
         return nil
       }
     }
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
     queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
 
     XCTAssertEqual(operation2.output, "10")
@@ -91,8 +88,7 @@ class InjectionTests: XCTestCase {
         return nil
       }
     }
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
     queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: true)
 
     XCTAssertNil(operation2.output)
@@ -109,8 +105,7 @@ class InjectionTests: XCTestCase {
 
     operation1.input = 10
     let adapterOperation = AdvancedOperation.injectOperation(operation1, into: operation2)
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
     queue.addOperations([operation1, operation2, adapterOperation, operation3], waitUntilFinished: false)
 
     waitForExpectations(timeout: 3)
@@ -127,8 +122,7 @@ class InjectionTests: XCTestCase {
     }
     operation1.input = 2000 // values greater than 1000 will produce a nil output 😎
     let adapterOperation = operation1.inject(into: operation2)
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
     queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: false)
 
     waitForExpectations(timeout: 3)
@@ -146,8 +140,7 @@ class InjectionTests: XCTestCase {
     }
 
     let adapterOperation = operation1.inject(into: operation2, requirements: [.successful])
-    let manager = ExclusivityManager()
-    let queue = AdvancedOperationQueue(exclusivityManager: manager)
+    let queue = AdvancedOperationQueue()
     queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: false)
 
     waitForExpectations(timeout: 3)
@@ -167,7 +160,7 @@ class InjectionTests: XCTestCase {
     XCTAssertTrue(operation1.isCancelled)
 
     let adapterOperation = operation1.inject(into: operation2, requirements: [.noCancellation])
-    let queue = AdvancedOperationQueue(exclusivityManager: ExclusivityManager())
+    let queue = AdvancedOperationQueue()
 
     queue.addOperations([operation1, operation2, adapterOperation], waitUntilFinished: false)